1.3-docker/HW2/stocks_products/logistic/views.py
from rest_framework.filters import SearchFilter
from rest_framework.viewsets import ModelViewSet

from logistic.models import Product, Stock
from logistic.serializers import ProductSerializer, StockSerializer


class ProductViewSet(ModelViewSet):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer
    filter_backends = [SearchFilter]
    search_fields = ['title', 'description']


class StockViewSet(ModelViewSet):
    queryset = Stock.objects.all()
    serializer_class = StockSerializer

    def get_queryset(self):
        queryset = super().get_queryset()
        search_str = self.request.query_params['products']
        return queryset.filter(products__title__icontains=search_str) or \
               queryset.filter(products__description__icontains=search_str)
