1.3-docker/HW2/stocks_products/requirements.txt
django
djangorestframework
django-filter
gunicorn
